+++ src/main/java/com/coastee/server/global/config/SecurityConfig.java
@@ -1,6 +1,6 @@
 package com.coastee.server.global.config;
 
-import com.coastee.server.login.util.JwtProvider;
+import com.coastee.server.jwt.util.JwtProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
