+++ src/main/java/com/coastee/server/jwt/util/JwtProvider.java
@@ -1,8 +1,9 @@
-package com.coastee.server.login.util;
+package com.coastee.server.jwt.util;
 
 import com.coastee.server.auth.domain.Authority;
 import com.coastee.server.global.apipayload.exception.handler.ExpiredPeriodJwtException;
 import com.coastee.server.global.apipayload.exception.handler.InvalidJwtException;
+import com.coastee.server.jwt.domain.AuthTokens;
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.security.Keys;
 import jakarta.annotation.PostConstruct;
@@ -34,6 +35,13 @@ public class JwtProvider {
     private void _getSecretKey() {
         String keyBase64Encoded = Base64.getEncoder().encodeToString(secretKey.getBytes());
         key = Keys.hmacShaKeyFor(keyBase64Encoded.getBytes());
+    }
+
+    public AuthTokens createTokens(final String subject) {
+        return AuthTokens.of(
+                createAccessToken(subject),
+                createRefreshToken()
+        );
     }
 
     public String createAccessToken(final String subject) {
